Add peg spawning functionality
