+++ models/square_peg/model.sdf
@@ -1,7 +1,7 @@
 <?xml version="1.0" ?>
 <sdf version="1.6">
     <model name='square_peg'>
-      <pose>0.264595 -0.617818 0.5 0 -0 0</pose>
+      <pose>0 0 0 1.57 0 0</pose>
       <link name='link'>
         <inertial>
           <mass>1</mass>
